Pause Removal bugifx attempt
Current theory is that the mp3 format is causing some hiccups with the
ffmpeg split funciton. The plan here is to have FreeSIWTCH record all
calls in wav and then perform the splits on the much more stable wav
format. After paus processing we can then convert it to mp3.